fixed display select for trips

+++ db/migrate/20180624123408_create_invoiced_trips.rb
@@ -16,9 +16,9 @@ class CreateInvoicedTrips < ActiveRecord::Migration[5.1]
       t.decimal  :uk_toll
       t.decimal  :netherlands_toll
       t.decimal  :km
-      t.decimal  :km_evogps
-      t.decimal  :km_driver_route_note
-      t.decimal  :total_amount
+      t.integer  :km_evogps
+      t.integer  :km_driver_route_note
+      t.integer  :total_amount
 
       t.timestamps
     end
